16 bit CLUT: point move bug
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@1960 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ CLUTOpacityView.mm
@@ -314,7 +314,7 @@
 		BOOL controlPointSelected = NO;
 		if([self isAnyPointSelected])
 		{
-			if((int) selectedPoint.x==(int) controlPoint.x && (int) selectedPoint.y==(int) controlPoint.y)
+			if((int) selectedPoint.x==(int) controlPoint.x && (float) selectedPoint.y==(float) controlPoint.y)
 			{
 				[selectedPointColor set];
 				controlPointSelected = YES;
@@ -354,7 +354,7 @@
 			BOOL selected = NO;
 			if([self isAnyPointSelected])
 			{
-				if((int) selectedPoint.x==(int) pt.x && (int) selectedPoint.y==(int) pt.y)
+				if((int) selectedPoint.x==(int) pt.x && (float) selectedPoint.y==(float) pt.y)
 				{
 					selected = YES;
 				}
@@ -478,7 +478,7 @@
 	for (i=0; i<[curves count] && curveIndex<0; i++)
 	{
 		NSPoint controlPoint = [self controlPointForCurveAtIndex:i];
-		if((int) selectedPoint.x==(int) controlPoint.x && (int) selectedPoint.y==(int) controlPoint.y)
+		if((int) selectedPoint.x==(int) controlPoint.x && (float) selectedPoint.y==(float) controlPoint.y)
 			curveIndex = i;
 	}
 	return curveIndex;
@@ -607,7 +607,7 @@
 			for (j=0; j<[aCurve count]; j++)
 			{
 				NSPoint pt = [[aCurve objectAtIndex:j] pointValue];
-				if((int) pt.x==(int) selectedPoint.x && (int) pt.y==(int) selectedPoint.y)
+				if((int) pt.x==(int) selectedPoint.x && (float) pt.y==(float) selectedPoint.y)
 				{
 					[self setColor:[[(NSColorPanel*)[notification object] color] colorUsingColorSpaceName: NSDeviceRGBColorSpace] forPointAtIndex:j inCurveAtIndex:i];
 					[self updateView];
@@ -615,7 +615,7 @@
 				}
 			}
 			NSPoint controlPoint = [self controlPointForCurveAtIndex:i];
-			if((int) controlPoint.x==(int) selectedPoint.x && (int) controlPoint.y==(int) selectedPoint.y)
+			if((int) controlPoint.x==(int) selectedPoint.x && (float) controlPoint.y==(float) selectedPoint.y)
 			{
 				[self setColor:[[(NSColorPanel*)[notification object] color] colorUsingColorSpaceName: NSDeviceRGBColorSpace] forCurveAtIndex:i];
 				[self updateView];
@@ -909,7 +909,7 @@
 				for (j=0; j<[aCurve count]; j++)
 				{
 					NSPoint pt = [[aCurve objectAtIndex:j] pointValue];
-					if((int) pt.x==(int) selectedPoint.x && (int) pt.y==(int) selectedPoint.y)
+					if((int) pt.x==(int) selectedPoint.x && (float) pt.y==(float) selectedPoint.y)
 					{
 						NSPoint newPoint = [transformView2Coordinate transformPoint:[self convertPoint:[theEvent locationInWindow] fromView:nil]];
 						newPoint = [self legalizePoint:newPoint inCurve:aCurve atIndex:j];
@@ -924,8 +924,8 @@
 			{	
 				firstPoint = [[aCurve objectAtIndex:0] pointValue];
 				lastPoint = [[aCurve lastObject] pointValue];
-				BOOL firstPointSelected = ((int) firstPoint.x==(int) selectedPoint.x && (int) firstPoint.y==(int) selectedPoint.y);
-				BOOL lastPointSelected = ((int) lastPoint.x==(int) selectedPoint.x && (int) lastPoint.y==(int) selectedPoint.y);
+				BOOL firstPointSelected = ((int) firstPoint.x==(int) selectedPoint.x && (float) firstPoint.y==(float) selectedPoint.y);
+				BOOL lastPointSelected = ((int) lastPoint.x==(int) selectedPoint.x && (float) lastPoint.y==(float) selectedPoint.y);
 				firstPoint = [transformCoordinate2View transformPoint:firstPoint];
 				lastPoint = [transformCoordinate2View transformPoint:lastPoint];
 				if( firstPointSelected || lastPointSelected)
@@ -962,7 +962,7 @@
 			}
 			
 			NSPoint controlPoint = [self controlPointForCurveAtIndex:i];
-			if( (int) controlPoint.x == (int) selectedPoint.x && (int) controlPoint.y == (int) selectedPoint.y)
+			if( (int) controlPoint.x == (int) selectedPoint.x && (float) controlPoint.y == (float) selectedPoint.y)
 			{			
 				NSPoint newPointInView = [self convertPoint:[theEvent locationInWindow] fromView:nil];
 				NSPoint newPoint = [transformView2Coordinate transformPoint:newPointInView];
@@ -1167,7 +1167,7 @@
 		for (j=0; j<[aCurve count] && curveIndex<0; j++)
 		{
 			NSPoint pt = [[aCurve objectAtIndex:j] pointValue];
-			if((int) selectedPoint.x==(int) pt.x && (int) selectedPoint.y==(int) pt.y)
+			if((int) selectedPoint.x==(int) pt.x && (float) selectedPoint.y==(float) pt.y)
 				curveIndex = i;
 		}
 	}
@@ -1177,7 +1177,7 @@
 		for (i=0; i<[curves count] && curveIndex<0; i++)
 		{
 			NSPoint controlPoint = [self controlPointForCurveAtIndex:i];
-			if((int) selectedPoint.x==(int) controlPoint.x && (int) selectedPoint.y==(int) controlPoint.y)
+			if((int) selectedPoint.x==(int) controlPoint.x && (float) selectedPoint.y==(float) controlPoint.y)
 				curveIndex = i;
 		}
 	}
@@ -1245,7 +1245,7 @@
 				for (j=0; j<[aCurve count]; j++)
 				{
 					NSPoint pt = [[aCurve objectAtIndex:j] pointValue];
-					if((int) selectedPoint.x==(int) pt.x && (int) selectedPoint.y==(int) pt.y)
+					if((int) selectedPoint.x==(int) pt.x && (float) selectedPoint.y==(float) pt.y)
 					{
 						NSData* colorData = [NSArchiver archivedDataWithRootObject:[[pointColors objectAtIndex:i] objectAtIndex:j]];
 						NSPasteboard* pasteboard = [NSPasteboard generalPasteboard];
@@ -1305,7 +1305,7 @@
 				for (j=0; j<[aCurve count]; j++)
 				{
 					NSPoint pt = [[aCurve objectAtIndex:j] pointValue];
-					if((int) selectedPoint.x==(int) pt.x && (int) selectedPoint.y==(int) pt.y)
+					if((int) selectedPoint.x==(int) pt.x && (float) selectedPoint.y==(float) pt.y)
 					{
 						NSData* colorData = [pasteboard dataForType:type];
 						NSColor *color = [NSUnarchiver unarchiveObjectWithData:colorData];
@@ -1338,7 +1338,7 @@
 				for (j=0; j<[aCurve count]; j++)
 				{
 					NSPoint pt = [[aCurve objectAtIndex:j] pointValue];
-					if((int) selectedPoint.x==(int) pt.x && (int) selectedPoint.y==(int) pt.y)
+					if((int) selectedPoint.x==(int) pt.x && (float) selectedPoint.y==(float) pt.y)
 					{
 						if([aCurve count]<=3)
 						{
